% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PoliticalDatasets.R
\name{count_sequence_breaks}
\alias{count_sequence_breaks}
\title{Utility function for counting sequence breaks}
\usage{
count_sequence_breaks(seq, seq_step = 1)
}
\arguments{
\item{seq}{}

\item{seq_step}{}
}
\value{
a vector of periods
}
\description{
Utility function for counting sequence breaks
}

