% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expanded_gdp_documentation.R
\docType{data}
\name{expanded_gdp}
\alias{expanded_gdp}
\title{Expanded GDP and Trade Data, compiled by Kristian Skrede Gleditsch, v. 6.0}
\format{A data frame with 9627 rows and 8 variables: \describe{
  \item{statenum}{Numeric country identifier} \item{statid}{Three letter
  acronym country identifier For a list of the country id and numeric codes,
  see \url{http://privatewww.essex.ac.uk/~ksg/statelist.html}. Further
  details are available in case description file and the paper describing the
  data: Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate System
  Membership: A Revised List of the Independent States since 1816."
  International Interactions 25: 393-413} \item{year}{year of observation}
  \item{pop}{population, see also notes below} \item{realgdp}{total real GDP,
  2005} \item{rgdppc}{real GDP per capita, 2005 prices} \item{cgdppc}{GDP per
  capita, current prices} \item{origin}{origin code for observation} }}
\source{
\url{http://privatewww.essex.ac.uk/~ksg/exptradegdp.html}
}
\usage{
expanded_gdp
}
\description{
These data provide estimates of GDP per capita of independent states. The
core input data used is PWT 8.0. Unlike previous versions of the GDP data
this reports PPP estimates of real GDP rather than per capita GDP.
}
\details{
These data are explained in greater detail in Gleditsch, Kristian S. 2002.
"Expanded Trade and GDP Data," Journal of Conflict Resolution 46: 712-24.
\url{http://jcr.sagepub.com/content/46/5/712.short}.

  \emph{Table of Origin Codes} \itemize{
  \item 0	- From PWT 8.0 (but see
  expections for population figures below)
  \item  -1	- From PWT 5.6
  \item -2	- From Maddison Project Database
  \item 1	- Imputations for lead/tails based on first/last available value,
  deflated to current value for gdppc using the GDP deflator from the Bureau
  of Economic Analysis
  \item 2	Interpolated value
  (within series)
  \item 3	Estimates based on figures from World Bank Global
  development indicators, using shares to reference countries. (See
  ksgmdw/scaling.asc for details)}

  \emph{Population figures: Origins and exceptions}

  a) All missing population figures in the PWT data have been replaced by
  population figures from the Expanded Population Data v.2.0, compiled by
  Kristian Skrede Gleditsch. This file is is provided as
  ksgmdw/countrypopestimates.asc, see the associated readme file for details
  on the origin codes. No documention is available as of yet for these data.
  These estimates are generally based on country specific censuses, and the
  particular references for each are available on request.

  b) The PWT populations estimates have been retained when available, with
  some exceptions

  The PWT project lists population figures for countries that experience
  large boundary changes based on the current size of the existing state.
  This means that the PWT population figures for the Federal Republic of
  Germany *before* the end of the German Democratic Republic reflect the
  population size of the two Germanies combined. Likewise, pre-1991
  population figures for the Soviet Union reflect the size of present day
  Russia, even though this is only about half the size of former Soviet
  Union. I am grateful to Ye Wang from the PWT project for explicity
  confirming this to me.

  Since the present size of a state seems inappropriate for historical
  comparisons, the current version of the data replaces the PWT figures with
  estimates from the Expanded Population Data v.2.0 in the following
  instances:

  A. Population of states prior to merging with other states i) Federal
  Republic of Germany through 1990 (prior to inclusion of former Democratic
  Republic of Germany) ii) Democratic Republic of Vietnam through 1975 (prior
  to inclusion of former Republic of (South) Vietnam iii) Arab Republic of
  (North) Yemen through 1990 (prior to merger with former People's Republic
  of (South) Yemen

  B. Population of states prior to secession of new states i) Pakistan
  through 1970 (prior to secession of Bangladesh) ii) Soviet Union through
  1990 (prior to the secession of multiple former Soviet Republics and the
  new Russian Federation) iii) Yugoslavia through 1991 (prior to the
  independence of Bosnia, Croatia, Slovenia, and Macedonia)
}
\keyword{datasets}

