% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PoliticalDatasets.R
\name{to_gw_system}
\alias{to_gw_system}
\title{Merge a dataset with country and date info with the Gleditsch-Ward system of states}
\usage{
to_gw_system(data, country_col = "country", date_col = "year",
  match_on = "overlap", include_extras = TRUE)
}
\arguments{
\item{data}{}

\item{country_col}{}

\item{date_col}{}
}
\value{
a dataset merged with the Gleditsch-Ward system of states \link{gw_system}
}
\description{
Merge a dataset with country and date info with the Gleditsch-Ward system of states
}
\examples{
library(dplyr)
data <- bmr \%>\% select(bmr_country, bmr_ccode, year, democracy) \%>\% filter(grepl("YUG|SERBIA",bmr_country))
to_gw_system(data, "bmr_country", "year")
}

